finish review, reorder and update readme
