--- lib/gopar/main.go
@@ -1,9 +1,0 @@
-package libgopar
-
-import (
-	"github.com/akalin/gopar/cmd/par"
-)
-
-func Gopar(args_ []string) int {
-	return (Par(args_))
-}
